@@ -6,30 +6,26 @@
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
-    "@fortawesome/react-fontawesome": "^0.1.9",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@xstate/react": "^1.6.1",
-    "chart.js": "^3.5.1",
+    "@fortawesome/react-fontawesome": "^0.1.17",
+    "@xstate/react": "^2.0.0",
+    "chart.js": "^3.7.0",
     "d3": "^7.1.0",
     "d3-cloud": "^1.2.5",
     "d3-selection": "^3.0.0",
-    "immer": "^9.0.2",
-    "randomcolor": "^0.5.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.2.0",
-    "typescript": "^4.0.0",
-    "xstate": "^4.13.0"
+    "immer": "^9.0.12",
+    "randomcolor": "^0.6.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "styled-components": "^5.3.3",
+    "typescript": "^4.5.5",
+    "xstate": "^4.29.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview",
+    "test": "jest",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx src",
     "check": "tsc && npm run lint",
     "lint-fix": "eslint --ext .js,.jsx,.ts,.tsx src --fix"
@@ -47,33 +43,41 @@
     ]
   },
   "devDependencies": {
-    "@types/chart.js": "^2.9.19",
-    "@types/d3": "^5.7.2",
-    "@types/d3-cloud": "^1.2.3",
-    "@types/jest": "^24.9.1",
-    "@types/node": "^12.0.0",
-    "@types/randomcolor": "^0.5.4",
-    "@types/react": "^16.9.32",
-    "@types/react-dom": "^16.9.6",
-    "@types/react-router-dom": "^5.1.6",
-    "@types/styled-components": "^5.1.4",
-    "@typescript-eslint/eslint-plugin": "^4.5.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-airbnb-typescript": "^14.0.0",
+    "@babel/core": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/chart.js": "^2.9.35",
+    "@types/d3": "^7.1.0",
+    "@types/d3-cloud": "^1.2.5",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.15",
+    "@types/randomcolor": "^0.5.6",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.21",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "babel-jest": "^27.5.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^16.1.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-unicorn": "^36.0.0",
-    "prettier": "^2.1.2"
-  },
-  "jest": {
-    "transformIgnorePatterns": [
-      "/node_modules/(?!(d3-selection))/"
-    ]
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^26.0.0",
+    "eslint-plugin-jest-dom": "^4.0.1",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-unicorn": "^40.1.0",
+    "jest": "^27.5.0",
+    "prettier": "^2.5.1",
+    "react-test-renderer": "^17.0.2",
+    "vite": "^2.7.13"
   }
 }
