@@ -1,5 +1,5 @@
 import React, { Suspense, lazy } from 'react';
-import { Route, Switch, useHistory } from 'react-router-dom';
+import { Route, Routes, useNavigate } from 'react-router-dom';
 import { Home } from './home';
 import { Spinner } from './common/molecules/Spinner';
 import { OrdskyHeader } from './common/molecules/Header';
@@ -12,10 +12,10 @@ import { About } from './about';
 const CollaborativePage = lazy(() => import('./collaborative'));
 
 const App: React.FC = function App() {
-  const history = useHistory();
+  const navigate = useNavigate();
 
-  const navigateToWordsPage = (): void => history.push('/words');
-  const navigateToTextPage = (): void => history.push('/text');
+  const navigateToWordsPage = (): void => navigate('/words');
+  const navigateToTextPage = (): void => navigate('/text');
 
   return (
     <>
@@ -25,23 +25,13 @@ const App: React.FC = function App() {
           <WordsProvider>
             <TextProvider>
               <Suspense fallback={<Spinner message="Laster side..." />}>
-                <Switch>
-                  <Route path="/words">
-                    <WordsPage onClickToTextForm={navigateToTextPage} />
-                  </Route>
-                  <Route path="/text">
-                    <TextPage onClickToWordsForm={navigateToWordsPage} />
-                  </Route>
-                  <Route path="/collab">
-                    <CollaborativePage />
-                  </Route>
-                  <Route path="/about">
-                    <About />
-                  </Route>
-                  <Route path="/">
-                    <Home onClickCreate={navigateToWordsPage} />
-                  </Route>
-                </Switch>
+                <Routes>
+                  <Route path="/words" element={<WordsPage onClickToTextForm={navigateToTextPage} />} />
+                  <Route path="/text" element={<TextPage onClickToWordsForm={navigateToWordsPage} />} />
+                  <Route path="/collab" element={<CollaborativePage />} />
+                  <Route path="/about" element={<About />} />
+                  <Route path="/" element={<Home onClickCreate={navigateToWordsPage} />} />
+                </Routes>
               </Suspense>
             </TextProvider>
           </WordsProvider>
